Move `SetOnce*` under `option` package
