+++ java/src/main/kotlin/io/spine/validation/java/option/setonce/SetOnceRenderer.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2024, TeamDev. All rights reserved.
+ * Copyright 2025, TeamDev. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -24,7 +24,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.validation.java.setonce
+package io.spine.validation.java.option.setonce
 
 import io.spine.protodata.ast.Field
 import io.spine.protodata.ast.PrimitiveType
@@ -43,7 +43,7 @@ import io.spine.string.shortly
 import io.spine.validation.SET_ONCE
 import io.spine.validation.SetOnceField
 import io.spine.validation.java.findMessageTypes
-import io.spine.validation.java.setonce.SetOnceNumberField.Companion.SupportedNumbers
+import io.spine.validation.java.option.setonce.SetOnceNumberField.Companion.SupportedNumbers
 
 /**
  * Takes the discovered [SetOnceField]s and modifies their Java builders to make sure
